Split off mount support functions to dedicated module
Note that unlike what was there before, all functions are always defined,
regardless of STRICT_CONFINEMENT. The body of some functions ends up being
empty but they are always there.
