+++ src/mount-support.h
@@ -1,0 +1,26 @@
+/*
+ * Copyright (C) 2015 Canonical Ltd
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 3 as
+ * published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+#ifndef SNAP_MOUNT_SUPPORT_H
+#define SNAP_MOUNT_SUPPORT_H
+
+void setup_private_mount(const char *appname);
+void setup_private_pts();
+void setup_snappy_os_mounts();
+void setup_slave_mount_namespace();
+
+#endif
